Renamed collection_util, moved glob_util.
