+++ st3/sublime_lib/_util/collections.py
